rust-rewrite: phase 3.2 clean code precedence
